// types.ts
export type InputEnvConfig = {
    btp_admin_email: string;
    btp_admin_password: string;
    btp_global_account_subdomain: string;
    use_existing_subaccount: string;
    btp_subaccount_domain: string;
    btp_subaccount_name: string;
    btp_subaccount_region: string;
    cloudfoundry_service_plan: string;
    cloudfoundry_org: string;  
    cloudfoundry_space: string;
    idp_origin: string;
}